==> 00 exploration, preprocessing.Rmd <==
---
title: "00 exploration, preprocessing"
author: "Brenna Kelly"
date: "2023-10-22"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r results='hide', message=FALSE, warning=FALSE}
library(boot)
library(purrr)
library(dplyr)
library(caret)
library(tidyr)
library(lmtest)
library(gifski)
library(moments)
library(ggplot2)
library(splines)
library(tableone)
library(gridExtra)
```
# Data Exploration and Preprocessing

### Summarizing variables, generating Table 1

```{r}

growth <- read.csv("utah_fetal_growth.csv")
names(growth) <- tolower(names(growth))

summary(growth)

# issues identified:
table(growth$race, growth$dhispanic) # since "other" is also unknown, "missing" will be imputed as "other"

growth <- growth |>
  mutate(bpreterm = ifelse(bpreterm == 1, "preterm", "not preterm")) |>
  mutate(t1_strata = case_when(dfemale == 1 & 
                                 gestational_age >= median(gestational_age) ~ "female, upper",
                               dfemale == 1 & 
                                 gestational_age < median(gestational_age) ~ "female, lower",
                               dfemale == 0 & 
                                 gestational_age >= median(gestational_age) ~ "male, upper",
                               dfemale == 0 & 
                                 gestational_age < median(gestational_age) ~ "male, lower")) |>
  mutate(race_eth = ifelse(dhispanic == "1", "Hispanic", race)) |>
  mutate(race_eth = ifelse(race_eth == "Missing", "Other", race_eth))

names(growth)

tab2 <- CreateTableOne(vars = c("birth_weight", "parity", "bmomage", 
             "mom_height_inches", "mom_weight_previous", "race_eth"),
             data = growth, strata = "t1_strata")

print(tab2, showAllLevels = TRUE)
write.csv(print(tab2, showAllLevels = TRUE), "table_1.1.csv", row.names = TRUE)

```

### Skewness tests and potential remedies

```{r, warning = FALSE, message = FALSE}

hist_vars <- c("bmomage",
               "bbthyear",
               "gestational_age",
               "parity",
               "mom_height_inches",
               "mom_weight_previous")

plot_fx <- function(i){
  ggplot(data = growth, aes(x = growth[, which(names(growth) == i)])) +
    geom_histogram(fill = "plum4", bins = 32) +
    theme_minimal() +
    labs(x = i, y = "Freq")
}

plots <- map(hist_vars, plot_fx)

grid.arrange(plots[[1]], plots[[2]], 
             plots[[3]], plots[[4]], 
             plots[[5]], plots[[6]])

skew_fx <- function(i){
  data.frame(cbind(i,
        skewness(growth[which(!is.na(growth[, i])), i])))
        #skewness(log(growth[which(!is.na(growth[, i])), i])))
}

skew_df <- do.call(rbind, lapply(hist_vars, FUN = skew_fx))
names(skew_df) <- c("variable", "skewness")
skew_df$skewness <- as.numeric(skew_df$skewness)
skew_df$interpretation <- case_when(abs(skew_df$skewness) < 0.5 ~ "approximately symmetric",
                                    abs(skew_df$skewness) >= 0.5 & abs(skew_df$skewness) < 1 ~ "moderately skewed",
                                    abs(skew_df$skewness) >= 1 ~ "highly skewed")
skew_df # results

# gestational age: would require a drastic and uninterpretable linear transformation
skewness(growth[, "gestational_age"]^(4))
skewness(growth[, "gestational_age"]^(5))

# parity: follows poisson distribution; linear transformation inappropriate
plots[[4]]

# mom weight previous: an adjusting variable only; interpretation will not be a concern
skewness(log(growth[, "mom_weight_previous"]), na.rm = TRUE)
skewness(1 / (growth[, "mom_weight_previous"])^(1/2), na.rm = TRUE) # approximately symmetric
growth$mom_wt_tx <- 1 / (growth$mom_weight_previous)^(1/2)
# will test both transformed and untransformed variable in lm for misbehavior

# race/ethnicity: creating an extra, simple variable
table(growth$race_eth)
prop.table(table(growth$race_eth))*100
# oversimplifying race not preferred
growth$race_eth_simple <- ifelse(growth$race_eth == "White", "White", "Non-White or Hispanic")
table(growth$race_eth_simple)
prop.table(table(growth$race_eth_simple))

```

### Imputation
Without imputation, we would be performing complete-case analysis, which is subject to bias. A full analysis may include a sensitivity analysis for imputation methods, but for our purposes, we'll just do some basic imputations. **Note:** we aren't performing imputation on our predictor variables or outcome variable.

```{r}

# it's probably reasonable to assume that if parity wasn't recorded, it was 0; this is something we can verify with the data contributors
growth$parity <- ifelse(is.na(growth$parity), 0, growth$parity)
growth$parity_cat <- as.factor(case_when(growth$parity == 0 ~ "0",
                                         growth$parity == 1 ~ "1",
                                         growth$parity == 2 ~ "2",
                                         growth$parity >= 3 ~ "3+",
                                         is.na(growth$parity) ~ "0")) # imputing 0

set.seed(30)
# random imputation for height and weight
random.imp <- function (a){
  missing <- is.na(a)
  n.missing <- sum(missing)
  a.obs <- a[!missing]
  imputed <- a
  imputed[missing] <- sample(a.obs, n.missing, replace = TRUE)
  return (imputed)
}

growth$mom_height_inches <- random.imp(growth$mom_height_inches)
growth$mom_weight_previous <- random.imp(growth$mom_weight_previous)

growth$bmomage <- ifelse(is.na(growth$bmomage), mean(growth$bmomage, na.rm = TRUE),
                         growth$bmomage)

```

### Transformations for interpretability
Centering, scaling, and setting referent categories.

```{r}

growth <- growth |>
  mutate(gestational_age_c = gestational_age - mean(gestational_age, na.rm = TRUE)) |>
  mutate(mom_wt_c_sc = (mom_weight_previous/10) - mean(mom_weight_previous, na.rm = TRUE)/10) |>
  mutate(mom_ht_c_sc = mom_height_inches/6 - mean((mom_height_inches/6), na.rm = TRUE)) |> # six-inch unit; but we won't need to interpret these variables anyways
  mutate(bmomage_c = bmomage - mean(bmomage, na.rm = TRUE)) |>
  mutate(bmi = (mom_weight_previous / (mom_height_inches^2)) * 703) |> # calculating bmi
  mutate(age_cat = case_when(bmomage < 24.5 ~ "18-24",
                             bmomage >= 24.5 & bmomage < 29.5 ~ "25-29",
                             bmomage >= 29.5 & bmomage < 34.5 ~ "30-34",
                             bmomage >= 34.5 ~ "35+")) |> # "geriatric" maternal age
  mutate(age_cat = relevel(as.factor(age_cat), ref = "18-24")) |> # ordinal referent; youngest adults
  mutate(parity_cat = relevel(parity_cat, ref = "0")) |> # logical referent
  mutate(parity_f = relevel(as.factor(parity), ref = "0")) |> # logical referent
  mutate(sex = relevel(as.factor(ifelse(dfemale == 1, "female", "male")), ref = "male")) |> # larger group
  mutate(race_eth = relevel(as.factor(race_eth), ref = "White")) |> # largest group
  mutate(race_eth_simple = relevel(as.factor(race_eth_simple), ref = "White")) |> # largest group
  mutate(bpreterm = as.factor(ifelse(is.na(bpreterm), "not preterm", "preterm"))) |>
  mutate(bpreterm = relevel(bpreterm, ref = "not preterm"))
  
# dataset for regression
growth_r <- growth |>
  select(-c(race, id, mom_height_inches, mom_wt_tx,
            mom_weight_previous,
            dblack, draceother, dasian, dhispanic, 
            dfemale, t1_strata))

summary(growth_r) # no missing values
names(growth)

#write.csv(growth_r, "growth_r.csv", row.names = FALSE)

```
### next step: testing assumptions
